+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Transitlog server changelog
 
+## Version 1.9.1, deployed 2020-06-22 @ 15:00
+
+- Update and fix stop events sort order.
+- Change de facto stop event from DEP to always be PDE with ODO.
+- Update virtual stop events for metro events and PDE event changes.
+- Fix alert urls.
+- Fix JORE objects dates-in-effect validation.
+- Fix week departure JORE objects validation with added day_type.
+
 ## Version 1.9.1, deployed 2020-04-16 @ 15:00
 
 - Fix route segments filter to remove invalid stops.